requirements.txt
requests
numpy
suno-bark
pydub
tqdm
python-dotenv
scipy
fastapi
uvicorn
jinja2
python-multipart
google-genai
edge-tts
moviepy>=2.0.0
Pillow
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
openpyxl
bytez
pydantic
schedule
